feat: Adicionando foto de um animal randômico na página de login
